On branch Add-Command-Parser  Your branch is ahead of 'origin/Add-Command-Parser' by 1 commit.    (use "git push" to publish your local commits)
 Changes to be committed:
	modified:   Unit_Tests_(VS_Code_Project)/.vscode/tasks.json
	new file:   Unit_Tests_(VS_Code_Project)/include/actions/Action_Interface.h
	new file:   Unit_Tests_(VS_Code_Project)/include/actions/Action_Updater.h
	renamed:    Unit_Tests_(VS_Code_Project)/include/Event_Scanner.h -> Unit_Tests_(VS_Code_Project)/include/input/Event_Scanner.h
	renamed:    Unit_Tests_(VS_Code_Project)/include/Input_Parser.h -> Unit_Tests_(VS_Code_Project)/include/input/Input_Parser.h
	renamed:    Unit_Tests_(VS_Code_Project)/include/KeyBindings.h -> Unit_Tests_(VS_Code_Project)/include/input/KeyBindings.h
	new file:   Unit_Tests_(VS_Code_Project)/src/actions/Action_Interface.cpp
	new file:   Unit_Tests_(VS_Code_Project)/src/actions/Action_Updater.cpp
	renamed:    Unit_Tests_(VS_Code_Project)/src/Event_Scanner.cpp -> Unit_Tests_(VS_Code_Project)/src/input/Event_Scanner.cpp
	renamed:    Unit_Tests_(VS_Code_Project)/src/Input_Parser.cpp -> Unit_Tests_(VS_Code_Project)/src/input/Input_Parser.cpp
	renamed:    Unit_Tests_(VS_Code_Project)/src/KeyBindings.cpp -> Unit_Tests_(VS_Code_Project)/src/input/KeyBindings.cpp
	modified:   Unit_Tests_(VS_Code_Project)/test_scripts/Learning/Command_Factory_learning.cpp
	modified:   Unit_Tests_(VS_Code_Project)/test_scripts/Learning/game_state_input_context.cpp
	modified:   Unit_Tests_(VS_Code_Project)/test_scripts/Learning/integrated_command_parser.cpp
	modified:   Unit_Tests_(VS_Code_Project)/test_scripts/Learning/simple_SDL_Input.cpp
	modified:   Unit_Tests_(VS_Code_Project)/test_scripts/test_Event_Scan_and_Key_Bindings.cpp
	modified:   Unit_Tests_(VS_Code_Project)/test_scripts/test_Event_Scanner.cpp
	modified:   Unit_Tests_(VS_Code_Project)/test_scripts/test_Input_Parser.cpp
-  Mjor Update!  Action_Interface and Action_Updater have been defined
   as their own separate header files.
-  Additionally, headers and .cpp files have been reorgainzed such that
   the main engine class is much smaller
