+++ Unit_Tests_(VS_Code_Project)/.vscode/tasks.json
@@ -17,10 +17,12 @@
 				"${workspaceFolder}\\src\\Mesh_Pipeline.cpp",
 				"${workspaceFolder}\\src\\Multiply_Matrix_Service.cpp",
 				"${workspaceFolder}\\src\\Triangle_Modifier.cpp",
-				"${workspaceFolder}\\src\\Event_Scanner.cpp",
-				"${workspaceFolder}\\src\\KeyBindings.cpp",		
+				"${workspaceFolder}\\src\\input\\Event_Scanner.cpp",
+				"${workspaceFolder}\\src\\input\\KeyBindings.cpp",		
 				"${workspaceFolder}\\src\\GameState_Observer_Pattern.cpp",			
-				"${workspaceFolder}\\src\\Input_Parser.cpp",
+				"${workspaceFolder}\\src\\input\\Input_Parser.cpp",
+				"${workspaceFolder}\\src\\actions\\Action_Interface.cpp",
+				"${workspaceFolder}\\src\\actions\\Action_Updater.cpp",
 //				"${workspaceFolder}\\src\\Rotator_Service.cpp",
 				"-I",
 				"C:\\SDL\\SDL2-2.0.18\\x86_64-w64-mingw32\\include",
