+++ Unit_Tests_(VS_Code_Project)/src/input/Input_Parser.cpp
@@ -1,4 +1,4 @@
-#include "Input_Parser.h"
+#include "input/Input_Parser.h"
 
 Input_Parser::Input_Parser(GameStateSubject &subject, Renderer &my_renderer, std::string binding_filename) :game_state_subject(subject){
     this->Engine_State=new Game_Engine_State_Observer(game_state_subject);
@@ -27,13 +27,15 @@ const std::unordered_map<std::string,float> Input_Parser::getRangeInput(){ retur
 void Input_Parser::scanInput(){
 
     input_events->scanInput();
-    if ((input_events->getTactileMap(PREVIOUS_MAP)!=input_events->getTactileMap(CURRENT_MAP))||(input_events->getRangeMap(PREVIOUS_MAP)!=input_events->getRangeMap(CURRENT_MAP))){
+    std::unordered_map<std::string, bool> curr_key_map = input_events->getTactileMap(CURRENT_MAP);
+    std::unordered_map<std::string, float> curr_range_input_map = input_events->getRangeMap(CURRENT_MAP);
+    std::unordered_map<std::string, bool> prev_key_map = input_events->getTactileMap(PREVIOUS_MAP);
+    std::unordered_map<std::string, float> prev_range_input_map = input_events->getRangeMap(PREVIOUS_MAP);    
+    std::cout << (curr_key_map!=prev_key_map) << "   " << (curr_range_input_map!=prev_range_input_map) << std::endl;
+    if ((curr_key_map!=prev_key_map)||(curr_range_input_map!=prev_range_input_map)){
         std::unordered_map<std::string, bool> curr_key_map = input_events->getTactileMap(CURRENT_MAP);
         std::unordered_map<std::string, float> curr_range_input_map = input_events->getRangeMap(CURRENT_MAP);
-        //std::unordered_map<std::string, bool> prev_key_map = input_events->getTactileMap(PREVIOUS_MAP);
-        //std::unordered_map<std::string, bool> curr_button_map = input_events->getTactileMap(MOUSE_BUTTON_MAP);
         std::unordered_map<std::string, bool> curr_command_map = bindings.getCommandMapFromKeycodes(curr_key_map);
-        //std::unordered_map<std::string, bool> prev_command_map = bindings.getCommandMapFromKeycodes(prev_key_map);
         this->curr_range_input = curr_range_input_map;
         this->curr_commands = curr_command_map;
